Task01 is done.

+++ C# OOP - Exercises/Polymorphism - Exercise/01.Vehicles/Contracts/IVehicle.cs
@@ -1,16 +1,11 @@
 ﻿namespace _01.Vehicles.Contracts;
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
 public interface IVehicle
 {
-    double FuelQuantity {  get; set; }
+    double FuelQuantity { get; set; }
     double FuelConsumption { get; init; }
 
-    void Drive(int distance, string className);
+    void Drive(double distance, string className);
+
     void Refuel(double fuelAmount);
 }
